Move menu items to the navigation package
This way we can reuse these components across workspaces (on profile for
 example).
