+++ app/MMB/src/components/header/Header.js
@@ -4,9 +4,9 @@ import * as React from 'react';
 import { View } from 'react-native';
 import { graphql, createFragmentContainer } from '@kiwicom/mobile-relay';
 import { StyleSheet, Color } from '@kiwicom/mobile-shared';
+import { SeparatorFullWidth } from '@kiwicom/mobile-navigation';
 import idx from 'idx';
 
-import { SeparatorFullWidth } from '../Separators';
 import StatusBar from './StatusBar';
 import TripInfo from './TripInfo';
 import HeaderImage from './HeaderImage';
