+++ app/MMB/src/menuGroups/TripInfoMenuGroup.js
@@ -3,9 +3,7 @@
 import * as React from 'react';
 import { TextIcon } from '@kiwicom/mobile-shared';
 import { Translation } from '@kiwicom/mobile-localization';
-
-import TitledMenuGroup from '../components/menu/TitledMenuGroup';
-import MenuItem from '../components/menu/MenuItem';
+import { MenuItem, TitledMenuGroup } from '@kiwicom/mobile-navigation';
 
 type Props = {|
   +openSubmenu: (activeId: string, menuId: string) => void,
