+++ packages/navigation/src/MenuGroup.js
@@ -3,7 +3,7 @@
 import * as React from 'react';
 import { View, Platform } from 'react-native';
 
-import { SeparatorFullWidth, SeparatorTrimmed } from '../Separators';
+import { SeparatorFullWidth, SeparatorTrimmed } from './Separators';
 
 type Props = {|
   children: React.Node,
