+++ packages/navigation/src/MenuItemIcon.js
